+++ managerClientPrometheus/lib/globalData/sensorObjects.py
@@ -1,0 +1,1 @@
+../../../shared_code/all/lib/globalData/sensorObjects.py